// src/models/RightDividendModel.ts
export interface RightDividendModel {
  sn: number
  company_name: string
  stock_symbol: string
  bonus: string
  cash: string
  total_dividend: string
  book_closure_date_ad: string
  book_closure_date_bs: string
  fiscal_year_ad: string
  fiscal_year_bs: string
  right_share: string
  right_book_close_date_ad: string
  right_book_close_date_bs: string
  status?: string
}